opt: reorganize execbuilder code
This change reorganizes the execbuilder code to mirror the
organization of the operators themselves:
 - rename `relational_builder.go` to `relational.go`
 - rename `scalar_builder.go` to `scalar.go`
 - pull out code for mutations to `mutation.go`
 - pull out code for non-DML statements to `statement.go`

Release note: None
